Fixed scout drone not rendering. Fixed bugs with tile destroy

+++ Assets/Scripts/Map/Tiles/MappedTiles/Tile.cs
@@ -163,12 +163,14 @@ namespace Amatsugu.Phos.Tiles
 			if(World.DefaultGameObjectInjectionWorld != null)
 			{
 				Map.EM.DestroyEntity(_tileEntity);
-				Map.EM.DestroyEntity(_decor);
+				if(_decor.IsCreated)
+					Map.EM.DestroyEntity(_decor);
 			}
 			if (_decor.IsCreated)
 			{
 				_decor.Dispose();
 			}
+			_isRendered = false;
 		}
 
 		public void Show(bool isShown)
